refactor(login): change directory name
