+++ src/pages/Checkout/index.js
@@ -1,8 +1,6 @@
-import Login from "../../components/login/login";
 const Checkout = () => {
-  return 
-  <>
-  </>;
+  return;
+  <></>;
 };
 
 export default Checkout;
